+++ Slim/Handlers/ExceptionHandler.php
@@ -15,13 +15,9 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Slim\Exception\HttpException;
 use Slim\Exception\HttpMethodNotAllowedException;
-use Slim\Interfaces\ContainerResolverInterface;
+use Slim\Interfaces\ContentNegotiatorInterface;
 use Slim\Interfaces\ExceptionHandlerInterface;
-use Slim\Interfaces\ExceptionRendererInterface;
 use Throwable;
-
-use function explode;
-use function strtolower;
 
 /**
  * This handler determines the response based on the media type (mime)
@@ -33,35 +29,32 @@ final class ExceptionHandler implements ExceptionHandlerInterface
 {
     private ResponseFactoryInterface $responseFactory;
 
-    private ContainerResolverInterface $resolver;
-
-    /* @var array<ExceptionHandlerInterface|callable|string> */
-    private array $renderers = [];
-
-    private string $defaultMediaType = 'application/json';
-
     private bool $displayErrorDetails = false;
+
+    private ContentNegotiatorInterface $contentNegotiator;
 
     public function __construct(
         ResponseFactoryInterface $responseFactory,
-        ContainerResolverInterface $resolver
+        ContentNegotiatorInterface $contentNegotiator
     ) {
-        $this->resolver = $resolver;
         $this->responseFactory = $responseFactory;
+        $this->contentNegotiator = $contentNegotiator;
     }
 
     public function __invoke(ServerRequestInterface $request, Throwable $exception): ResponseInterface
     {
         $statusCode = $this->determineStatusCode($request, $exception);
-        $mediaType = $this->determineMediaType($request);
-        $response = $this->createResponse($statusCode, $mediaType, $exception);
-        $renderer = $this->determineRenderer($mediaType);
+        $negotiationResult = $this->contentNegotiator->negotiate($request);
+        $response = $this->createResponse($statusCode, $negotiationResult->getMediaType(), $exception);
 
-        // Invoke the renderer
-        /** @var ResponseInterface $response */
-        $response = call_user_func($renderer, $request, $response, $exception, $this->displayErrorDetails);
-
-        return $response;
+        // Invoke the formatter
+        return call_user_func(
+            $negotiationResult->getFormatter(),
+            $request,
+            $response,
+            $exception,
+            $this->displayErrorDetails
+        );
     }
 
     public function setDisplayErrorDetails(bool $displayErrorDetails): self
@@ -69,63 +62,6 @@ final class ExceptionHandler implements ExceptionHandlerInterface
         $this->displayErrorDetails = $displayErrorDetails;
 
         return $this;
-    }
-
-    public function registerRenderer(string $mediaType, ExceptionRendererInterface|callable|string $handler): self
-    {
-        $this->renderers[$mediaType] = $handler;
-
-        return $this;
-    }
-
-    public function setDefaultMediaType(string $mediaType): self
-    {
-        $this->defaultMediaType = $mediaType;
-
-        return $this;
-    }
-
-    /**
-     * Determine which renderer to use based on media type.
-     */
-    private function determineRenderer(string $mediaType): ExceptionRendererInterface
-    {
-        $renderer = $this->renderers[$mediaType] ?? $this->renderers[$this->defaultMediaType];
-
-        return $this->resolver->resolveCallable($renderer);
-    }
-
-    /**
-     * Determine which content type we know about is wanted Accept header.
-     *
-     * https://www.iana.org/assignments/media-types/media-types.xhtml
-     */
-    protected function determineMediaType(ServerRequestInterface $request): string
-    {
-        $mediaTypes = $this->parseAcceptHeader($request->getHeaderLine('Accept'));
-
-        if (!$mediaTypes) {
-            $mediaTypes = $this->parseContentType($request->getHeaderLine('Content-Type'));
-        }
-
-        // Use the order of definitions
-        foreach ($this->renderers as $mediaType => $_) {
-            if (isset($mediaTypes[$mediaType])) {
-                return $mediaType;
-            }
-        }
-
-        // No direct match is found. Check for +json or +xml.
-        foreach ($mediaTypes as $type) {
-            if (preg_match('/\+(json|xml)/', $type, $matches)) {
-                $mediaType = 'application/' . $matches[1];
-                if (isset($this->renderers[$mediaType])) {
-                    return $mediaType;
-                }
-            }
-        }
-
-        return $this->defaultMediaType;
     }
 
     private function determineStatusCode(ServerRequestInterface $request, Throwable $exception): int
@@ -156,34 +92,5 @@ final class ExceptionHandler implements ExceptionHandlerInterface
         }
 
         return $response;
-    }
-
-    public function parseAcceptHeader(string $accept = null): array
-    {
-        $acceptTypes = $accept ? explode(',', $accept) : [];
-
-        // Normalize types
-        $cleanTypes = [];
-        foreach ($acceptTypes as $type) {
-            $tokens = explode(';', $type);
-            $name = trim(strtolower(reset($tokens)));
-            $cleanTypes[$name] = $name;
-        }
-
-        return $cleanTypes;
-    }
-
-    private function parseContentType(string $contentType = null): array
-    {
-        $parts = explode(';', $contentType ?? '');
-
-        // @phpstan-ignore-next-line
-        if (!$parts) {
-            return [];
-        }
-
-        $name = strtolower(trim($parts[0]));
-
-        return [$name => $name];
     }
 }